mocking/countdown_test.go
package main

import (
	"reflect"
	"bytes"
	"testing"
)
const write = "write"
const sleep = "sleep"
type SpySleeper struct {
	Calls int
}
func (s *SpySleeper) Sleep() {
	s.Calls++
}
type CountdownOperationsSpy struct {
	Calls []string
}
func (s *CountdownOperationsSpy) Sleep()  {
	s.Calls = append(s.Calls, sleep)
}
func (s *CountdownOperationsSpy) Write(p []byte) (n int, err error) {
	s.Calls = append(s.Calls, write)
	return
}

func TestCountdown(t *testing.T) {
	t.Run("Get correct output", func (t *testing.T) {
	buffer := &bytes.Buffer{}
	spySleeper := &CountdownOperationsSpy{}
	Countdown(buffer, spySleeper)
	got := buffer.String()
	want := `3
2
1
Go!`

	if got != want {
		t.Errorf("got %q want %q", got, want)
	}
	})
	t.Run("sleep before every print", func (t *testing.T) {
		spySleepPrinter := &CountdownOperationsSpy{}
		Countdown(spySleepPrinter, spySleepPrinter)
		want := []string{
			sleep,
			write, 
			sleep,
			write,
			sleep,
			write,
			sleep,
			write,
		}
		if !reflect.DeepEqual(want, spySleepPrinter.Calls) {
			t.Errorf("wanted calls %v got %v", want, spySleepPrinter.Calls)
		}
	})
}